Added new functionality to action
Action will check the new action-fail-level parameter. If the vaule is equal to default it will work as previous version and only create json file. If the value is different then the default action-fail-level, action after creating json file will check if in the json file contain any annotation level that is equal to annotation levels listed in action-fail-level parameter. if comparison is true action will end with exit code -1 (this will create failure status for step running this action)